Made OS Check Opt-Out, Improved Prompt.bws
I personally comment out os checks, as I now develop this on Linux. I could understand other people that would want to run this not on Windows as well, if they wish. It is still very much not recommended to use this on Linux or Mac, as the real Bash is wayyyyy more ideal. But if you want to get this running on other operating systems, you can by adding "disableOSCheck" to prompt.bws and setting it to a value of 1. It will also come this way if you generate it from the repair script.

While adding this functionality, I also redeveloped the way the prompt.bws file is read, using Python's split function instead of the mess I made with loops inside loops. Similar to the new way a command is read and arguments are made.