New Installer
This new installer solves a problem I've always had while releasing new versions of Bash for Windows. The old installer used to work that it will write every script out line by line, so to release a new version, I have to rewrite every line of code in the installer script. This script makes things much easier. It will instead download the Startup.exe file and a Zip file of this repository from GitHub in a release I make. So now to make a new release I just have to draft a new release with the new Startup binary when I'm ready for a stable release, then put the new URLs in the new installer.

The old installers are now considered deprecated.